requirements.txt
--pre
--only-binary=all
numpy==1.24.4
pandas==2.0.3
pydantic==2.5.2
openai==1.6.1
gradio==4.44.0
python-dotenv==1.0.1
typer==0.12.3
rich==13.7.1